create service for load calculation
